Eliminate use of mktemp()
The use of mktemp() results in an unsuppressable warning message
on certain compilers, including gcc.  This causes problems for
projects which have rules regarding warning-free compiles.

Use mkstemp instead. The problem with mkstemp is that it opens the
file instead of just returning a unique file name.  However, as a
side effect it does change the template to a unique file name.
This can be used similarly to the return value from mktemp.