Fix CMake shared linker options on macOS.
The linker on macOS does not recognize the used linker options, so use different options that should have the same effect.
The exported_symbols_list option does not work with the names in matio.sym, it requires symbol names starting with '_', so create a copy of matio.sym and modify it when necessary.
Symbol names starting with '_' also works with the GNU linker but not with the LLVM Linker on Linux, so don't modify the original file.