Fix LFS support on MinGW 64bit builds
The primary change is to drop the
!defined(HAVE_FSEEKO) && !defined(HAVE_FTELLO) from the corresponding
condition, as MinGW defines both fseeko and fseeko64, and it doesn't
really make sense to not always use the 64bit versions when available.

The other change is that the '#define mat_off_t __int64' is moved into
the inner conditions of the Windows check, so that if no inner
condition is applicable, the later default definition in the header
doesn't clash with the previous definition.