Windows: don't check for MSVC for supporting Unicode filenames
The corresponding wchar_t functionality is also supported by
MinGW (at least somewhat recent versions thereof), so the check
for _MSC_VER causes the MinGW build to only support the local
codepage, and thus differ from the MSVC build, and also have
the issue of not being able to access some files on the filesystem
if they are not representable in the local codepage.

This drops the defined(_MSC_VER) and only keeps the _WIN32
check, because this is not a compiler feature, but rather a
platform feature, so that check is the correct one.